refactor(VisualiseForm) remove unused props and move PreviewPeriodPicker
