+++ ui/src/containers/Visualisations/components/PreviewPeriodPicker.js
